EOS-20944-SasPort & EOS-20956-NWport HealthInfo: moved the common template functions to ResourceMap. minor changes.
